refactor(CadastroObras): refatora componente de obras
